requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2.1
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
python-dotenv~=1.1.0
requests~=2.32.3
jsonrpcclient~=4.0.3